requirements.txt
anyio==3.6.1
appnope==0.1.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
command-not-found==0.3
contourpy==1.0.5
cryptography==38.0.1
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.1
django-environ==0.9.0
duplicity==0.8.12.0
entrypoints==0.4
environ==1.0
executing==1.1.1
fasteners==0.14.1
fastjsonschema==2.16.2
fonttools==4.37.4
fpdf==1.7.2
future==0.18.2
html5lib==1.0.1
httplib2==0.14.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.4.2
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterthemes==0.20.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lesscpy==0.15.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.5
more-itertools==4.2.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.10.4
nltk==3.7
notebook==6.5.1
notebook-shim==0.2.0
numpy==1.23.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
patsy==0.5.3
pdfminer.six==20220524
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
plotly==5.10.0
ply==3.11
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.6.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.15.0
pycups==1.9.73
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
PyPDF2==2.3.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.9.1
pytz==2022.4
pyxdg==0.26
PyYAML==5.3.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2022.9.13
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
scikit-learn==1.1.1
scipy==1.9.2
seaborn==0.12.1
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
statsmodels==0.13.2
striprtf==0.0.21
systemd-python==234
tenacity==8.0.1
terminado==0.16.0
testpath==0.4.4
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing-extensions==4.4.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.3
xkit==0.0.0
zipp==3.10.0
